Replace files with gRPC codes.
