--- server/health_grpc_test.go
@@ -1,131 +1,0 @@
-package server
-
-import (
-	"context"
-	"fmt"
-	"net/http"
-	"net/http/httptest"
-	"net/url"
-	"strconv"
-	"strings"
-	"time"
-
-	"github.com/cybozu-go/moco"
-	"github.com/cybozu-go/moco-agent/metrics"
-	"github.com/cybozu-go/moco-agent/test_utils"
-	"github.com/cybozu-go/moco/accessor"
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-	"github.com/prometheus/client_golang/prometheus"
-	healthpb "google.golang.org/grpc/health/grpc_health_v1"
-)
-
-func testHealthgRPC() {
-	var agent *Agent
-	var registry *prometheus.Registry
-	var gsrv healthpb.HealthServer
-
-	BeforeEach(func() {
-		test_utils.StopAndRemoveMySQLD(donorHost)
-		test_utils.StopAndRemoveMySQLD(replicaHost)
-
-		err := test_utils.StartMySQLD(donorHost, donorPort, donorServerID)
-		Expect(err).ShouldNot(HaveOccurred())
-		err = test_utils.StartMySQLD(replicaHost, replicaPort, replicaServerID)
-		Expect(err).ShouldNot(HaveOccurred())
-
-		err = test_utils.InitializeMySQL(donorPort)
-		Expect(err).ShouldNot(HaveOccurred())
-		err = test_utils.InitializeMySQL(replicaPort)
-		Expect(err).ShouldNot(HaveOccurred())
-
-		err = test_utils.PrepareTestData(donorPort)
-		Expect(err).ShouldNot(HaveOccurred())
-
-		err = test_utils.SetValidDonorList(replicaPort, donorHost, donorPort)
-		Expect(err).ShouldNot(HaveOccurred())
-
-		err = test_utils.ResetMaster(donorPort)
-		Expect(err).ShouldNot(HaveOccurred())
-		err = test_utils.ResetMaster(replicaPort)
-		Expect(err).ShouldNot(HaveOccurred())
-
-		registry = prometheus.NewRegistry()
-		metrics.RegisterMetrics(registry)
-
-		agent = New(test_utils.Host, clusterName, token, test_utils.MiscUserPassword, test_utils.CloneDonorUserPassword, replicationSourceSecretPath, "", replicaPort,
-			&accessor.MySQLAccessorConfig{
-				ConnMaxLifeTime:   30 * time.Minute,
-				ConnectionTimeout: 3 * time.Second,
-				ReadTimeout:       30 * time.Second,
-			},
-		)
-
-		gsrv = NewHealthService(agent)
-	})
-
-	AfterEach(func() {
-		test_utils.StopAndRemoveMySQLD(donorHost)
-		test_utils.StopAndRemoveMySQLD(replicaHost)
-	})
-
-	It("should return OK=true if no errors and cloning is not in progress", func() {
-		By("getting health")
-		res, err := gsrv.Check(context.Background(), &healthpb.HealthCheckRequest{})
-		Expect(err).ShouldNot(HaveOccurred())
-		Expect(res.Status).Should(Equal(healthpb.HealthCheckResponse_SERVING))
-	})
-
-	It("should return IsUnderCloning=true if cloning process is in progress", func() {
-		By("executing cloning")
-		err := test_utils.ResetMaster(replicaPort)
-		Expect(err).ShouldNot(HaveOccurred())
-		err = test_utils.SetValidDonorList(replicaPort, donorHost, donorPort)
-		Expect(err).ShouldNot(HaveOccurred())
-
-		req := httptest.NewRequest("GET", "http://"+replicaHost+"/clone", nil)
-		queries := url.Values{
-			moco.CloneParamDonorHostName: []string{donorHost},
-			moco.CloneParamDonorPort:     []string{strconv.Itoa(donorPort)},
-			moco.AgentTokenParam:         []string{token},
-		}
-		req.URL.RawQuery = queries.Encode()
-
-		res := httptest.NewRecorder()
-		agent.Clone(res, req)
-		Expect(res).Should(HaveHTTPStatus(http.StatusOK))
-
-		By("getting health expecting IsUnderCloning=true")
-		Eventually(func() error {
-			res, err := gsrv.Check(context.Background(), &healthpb.HealthCheckRequest{})
-			if res.Status == healthpb.HealthCheckResponse_NOT_SERVING && strings.HasSuffix(err.Error(), "isOutOfSynced=false, isUnderCloning=true") {
-				return nil
-			}
-			return fmt.Errorf("should become NOT_SERVING and IsUnderCloning=true: res=%s", res.String())
-		}, 5*time.Second, 200*time.Millisecond).Should(Succeed())
-
-		By("wating cloning is completed")
-		Eventually(func() error {
-			res, err := gsrv.Check(context.Background(), &healthpb.HealthCheckRequest{})
-			if err == nil {
-				return nil
-			}
-			return fmt.Errorf("should return without error: res=%s", res.String())
-		}, 30*time.Second, time.Second).Should(Succeed())
-	})
-
-	It("should return IsOutOfSynced=true if replica status has IO error", func() {
-		By("executing START SLAVE with invalid parameters")
-		err := test_utils.StartSlaveWithInvalidSettings(replicaPort)
-		Expect(err).ShouldNot(HaveOccurred())
-
-		By("getting health expecting IsOutOfSync=true")
-		Eventually(func() error {
-			res, err := gsrv.Check(context.Background(), &healthpb.HealthCheckRequest{})
-			if res.Status == healthpb.HealthCheckResponse_NOT_SERVING && strings.HasSuffix(err.Error(), "isOutOfSynced=true, isUnderCloning=false") {
-				return nil
-			}
-			return fmt.Errorf("should become NOT_SERVING and IsOutOfSynced=true: res=%s", res.String())
-		}, 5*time.Second, 200*time.Millisecond).Should(Succeed())
-	})
-}
